requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Cryptographie
cryptography==41.0.7
argon2-cffi==23.1.0

# Utilitaires
python-dateutil==2.8.2
uuid==1.30
pathlib==1.0.1

# SSL/TLS
pyOpenSSL==23.3.0

# Gestion des fichiers
python-magic==0.4.27

# Sécurité
bcrypt==4.1.2

# JSON
jsonschema==4.21.1
